Restructure tests and temporarily disable packaging.
