+++ compose/tests/test_docker_compose.py
