+++ core/tests/test_core.py
