+++ core/tests/test_docker_client.py
