+++ localstack/tests/test_localstack.py
