+++ rabbitmq/tests/test_rabbitmq.py
